added PageByPage, FitPage
There are situations where text of each page is given,
and it has to fit into the page.
PageByPage, FitPage objects are made to solve that situation.
page_by_page.txt contains text for each page of the the,
starting with page number as <p10>.
PageByPage parses the text into page folder with 0001, 0002 and so on...
FitPage laychs out each page,  and tries to fit content into a page by various tricks,
such as increasing font_size by increment of 0.2 until the last line extra space
is within acceptable lenth, and then re_aligning to fit the line.